LacosAPI/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
PyJWT==2.8.0
requests==2.31.0
bcrypt==4.0.1